Minor fix to make erlguten.app.src valid Erlang
By changing the placeholder [%MODULES%] to ["%MODULES%"] and adapting
the Perl script accordingly, erlguten.app.src can be turned into valid
Erlang syntax. This allows it to be processed via file:consult and
therefore rebar.

With this fix in place, depending on erlguten in rebar.config
is as simple as:
{deps,[{erlguten, ".*", {git, "https://github.com/richcarl/erlguten"}}]}.
{post_hooks,[{'get-deps', 'make -C deps/erlguten'}]}.